fix issue #I3YVBH 兼容oracle 11g add load all users/roles for flowinstance fix issue #I3W5YR

+++ OpenAuth.Repository/Core/LongEntity.cs
@@ -10,7 +10,7 @@ namespace OpenAuth.Repository.Core
     public class LongEntity :BaseEntity
     {
         [Browsable(false)]
-        public long Id { get; set; }
+        public decimal Id { get; set; }
         public override bool KeyIsNull()
         {
             return Id == 0;
